requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==2.12.0
deprecation==2.1.0
distro==1.9.0
fastapi==0.116.1
ffmpeg-python==0.2.0
frozenlist==1.7.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
openai==1.97.1
packaging==25.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
